Contract process done

+++ contract-paypal/src/model/services/OlinePaymentService.java
@@ -2,5 +2,5 @@ package model.services;
 
 public interface OlinePaymentService {
     Double paymentFee(Double amount);
-    Double insterest(Double amount, Double months);
+    Double insterest(Double amount, Integer months);
 }
